% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f.machine_learning.R
\name{cluster_samples}
\alias{cluster_samples}
\title{Cluster Samples}
\usage{
cluster_samples(data)
}
\arguments{
\item{data}{A data frame with genes as rows and samples as columns.}
}
\description{
Clusters the samples based on gene expression profiles.
}
