% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f.statistical_analysis.R
\name{correlation_matrix}
\alias{correlation_matrix}
\title{Compute Correlation Matrix}
\usage{
correlation_matrix(data)
}
\arguments{
\item{data}{A data frame with genes as rows and samples as columns.}
}
\value{
A correlation matrix.
}
\description{
Calculates a correlation matrix for gene expression data.
}
