% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f.statistical_analysis.R
\name{differential_expression}
\alias{differential_expression}
\title{Differential Expression Analysis}
\usage{
differential_expression(data, group_labels)
}
\arguments{
\item{data}{A data frame with genes as rows and samples as columns.}

\item{group_labels}{A factor indicating group labels for each sample.}
}
\value{
A data frame with log fold changes and p-values for each gene.
}
\description{
Conducts a basic differential expression analysis.
}
