% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f.machine_learning.R
\name{feature_selection}
\alias{feature_selection}
\title{Feature Selection}
\usage{
feature_selection(data, labels)
}
\arguments{
\item{data}{A data frame with genes as rows and samples as columns.}

\item{labels}{A factor with class labels for each sample.}
}
\description{
Performs feature selection to identify significant genes.
}
