% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f.functional_enrichment.R
\name{gene_ontology_enrichment}
\alias{gene_ontology_enrichment}
\title{Gene Ontology Enrichment}
\usage{
gene_ontology_enrichment(gene_list)
}
\arguments{
\item{gene_list}{A character vector of gene identifiers.}
}
\value{
A data frame with GO terms and enrichment scores.
}
\description{
Performs gene ontology enrichment analysis.
}
