% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f.network_analysis.R
\name{get_ppi_network}
\alias{get_ppi_network}
\title{Retrieve Protein-Protein Interaction Network}
\usage{
get_ppi_network(gene_list)
}
\arguments{
\item{gene_list}{A character vector of gene identifiers.}
}
\description{
Retrieves a PPI network for a given gene list.
}
