% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f.group_comparison_stats.R
\name{Group_comparison}
\alias{Group_comparison}
\title{Compare Biological Groups with t-test or ANOVA}
\usage{
Group_comparison(data, group, test_type = "t-test")
}
\arguments{
\item{data}{A data frame.}

\item{group}{A grouping variable.}

\item{test_type}{"t-test" or "anova".}
}
\value{
Results
}
\description{
Compare Biological Groups with t-test or ANOVA
}
