% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f.visualizations.R
\name{heatmap_gene_expression}
\alias{heatmap_gene_expression}
\title{Heatmap of Gene Expression}
\usage{
heatmap_gene_expression(data)
}
\arguments{
\item{data}{A data frame with genes as rows and samples as columns.}
}
\description{
Plots a heatmap for gene expression data.
}
