% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f.data_import.R
\name{import_data}
\alias{import_data}
\title{Import Biological Data}
\usage{
import_data(file_path, file_type = "csv")
}
\arguments{
\item{file_path}{Path to the data file.}

\item{file_type}{Type of file, e.g., "csv" or "excel".}
}
\value{
A data frame with imported data.
}
\description{
Import Biological Data
}
