% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f.data_preprocessing.R
\name{log_transform}
\alias{log_transform}
\title{Log Transform Data}
\usage{
log_transform(data)
}
\arguments{
\item{data}{A data frame with genes as rows and samples as columns.}
}
\value{
A log-transformed data frame.
}
\description{
Performs log transformation on gene expression data.
}
