% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f.functional_enrichment.R
\name{motif_enrichment}
\alias{motif_enrichment}
\title{Motif Enrichment Analysis}
\usage{
motif_enrichment(gene_list)
}
\arguments{
\item{gene_list}{A character vector of gene identifiers.}
}
\value{
A data frame with motifs and enrichment scores.
}
\description{
Detects over-represented motifs in the provided gene list.
}
