% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f.data_preprocessing.R
\name{normalize_counts}
\alias{normalize_counts}
\title{Normalize Counts}
\usage{
normalize_counts(data)
}
\arguments{
\item{data}{A data frame with genes as rows and samples as columns.}
}
\value{
A normalized data frame.
}
\description{
Applies a simple normalization technique to make counts comparable across samples.
}
