% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f.normalize_expression.R
\name{normalize_expression}
\alias{normalize_expression}
\title{Normalize Gene/Protein Expression}
\usage{
normalize_expression(data, method = "log")
}
\arguments{
\item{data}{A data frame with expression values.}

\item{method}{Method of normalization, e.g., "log" or "z-score".}
}
\value{
A normalized data frame.
}
\description{
Normalize Gene/Protein Expression
}
