% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f.statistical_analysis.R
\name{p_adjust}
\alias{p_adjust}
\title{Adjust p-values}
\usage{
p_adjust(p_values)
}
\arguments{
\item{p_values}{A numeric vector of p-values.}
}
\value{
A numeric vector of adjusted p-values.
}
\description{
Adjusts p-values for multiple testing.
}
