% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f.functional_enrichment.R
\name{pathway_enrichment}
\alias{pathway_enrichment}
\title{Pathway Enrichment Analysis}
\usage{
pathway_enrichment(gene_list)
}
\arguments{
\item{gene_list}{A character vector of gene identifiers.}
}
\value{
A data frame with enriched pathways and p-values.
}
\description{
Performs pathway enrichment analysis for a list of genes.
}
