% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f.expression_heatmap.R
\name{plot_expression_heatmap}
\alias{plot_expression_heatmap}
\title{Plot Expression Heatmap}
\usage{
plot_expression_heatmap(data)
}
\arguments{
\item{data}{A data frame of expression data.}
}
\value{
A heatmap
}
\description{
Plot Expression Heatmap
}
