% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f.visualizations.R
\name{plot_ma}
\alias{plot_ma}
\title{MA Plot}
\usage{
plot_ma(data)
}
\arguments{
\item{data}{A data frame with columns for average expression and log fold change.}
}
\description{
Creates an MA plot for gene expression data. Log ratio (M) and log average (A).
}
