% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f.visualizations.R
\name{plot_pca}
\alias{plot_pca}
\title{PCA Plot}
\usage{
plot_pca(data)
}
\arguments{
\item{data}{A data frame with genes as rows and samples as columns.}
}
\description{
Generates a PCA plot for sample clustering (Principal component analysis).
}
