% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f.visualizations.R
\name{plot_volcano}
\alias{plot_volcano}
\title{Volcano Plot}
\usage{
plot_volcano(data)
}
\arguments{
\item{data}{A data frame with columns for log fold change and p-value.}
}
\description{
Creates a volcano plot for differential expression results.
}
