% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f.machine_learning.R
\name{predict_biomarkers}
\alias{predict_biomarkers}
\title{Predict Biomarkers}
\usage{
predict_biomarkers(data, labels)
}
\arguments{
\item{data}{A data frame with genes as rows and samples as columns.}

\item{labels}{A factor with class labels for each sample.}
}
\description{
Uses "machine learning" to predict biomarkers from gene expression data.
}
