% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f.network_analysis.R
\name{visualize_network}
\alias{visualize_network}
\title{Visualize Network}
\usage{
visualize_network(network)
}
\arguments{
\item{network}{An igraph network object.}
}
\description{
Visualizes a network of interactions.
}
